restructured utils
